feat: 코드 효율성 증가를 위한 변경
